requirements.txt
python-dotenv
openai
langchain
streamlit
pydantic
pytest
